Add script to delete older entries
A script to delete older entries was created.

The purpose of the task was to create a TTL for the
database entries. Since Postgres does not support it there
were two alternatives:

1. Triggers:
	- This would probably require a trigger on insert, which
	  I feel is not only innefficient but also overkill.
2. A script associated to a sheduler job:
	- Easier to write and iterate.
	- But the deployment environment may not running scripts
 	  or have a scheduler.

The latter was chosen for ease of development. Both Heroku
and MEza suppor scheduling of jobs, so we have no practical
restrictions.

Part of the work of this job was make with the help of the
research made by Valter Nepomuceno on the context of
TEL-614 [PR](https://github.com/Talkdesk/legacy-translation/pull/54)