Upgrade Docker image,  Ruby, Bundler, pg version.
Docker image was upgraded to latest, in the "slim" version.

Ruby and Bundler were upgraded to latest versions, 2.6.3
and 2.0.1 respectively.

Pg was upgraded to 1.1.4 because 1.1.3 did not support
ruby 2.6. `libpq-dev` was added to runtime dependencies
because this gem need it. The old docker image wasn't
"slim" so it probably already had it installed.

Gemfile.lock was re-generated.