Tracks ongoing and last 24 hour calls
The app will show on `/` the number of currently ongoing
calls and the number of calls the ocurred in the last 24
hours.

It is currently using the date in which the first event for
the interaction was received. This means that an event is,
for some reason, delayed it will be considered as ocurring
later.

This also assumes that the events are received in the
correct order.

This concerns may be addressed in a future development but,
given the context of the app, I believe it is good enough
as it is.